requirements.txt
anyio==4.2.0
certifi==2023.11.17
cffi==1.16.0
cryptography==42.0.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==1.26.3
pandas==2.2.0
pycparser==2.21
PyMySQL==1.1.0
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
tzdata==2023.4
